Abstract away extractItem from extract to pave the way to the browser
